IGNITE-22733 Introduce TransactionChanges and abstract transactional test
