+++ modules/calcite/src/main/java/org/apache/ignite/internal/processors/query/calcite/exec/IndexScan.java
@@ -19,7 +19,6 @@ package org.apache.ignite.internal.processors.query.calcite.exec;
 import java.lang.reflect.Type;
 import java.util.Iterator;
 import java.util.List;
-import java.util.Set;
 import org.apache.calcite.rel.type.RelDataType;
 import org.apache.calcite.util.ImmutableBitSet;
 import org.apache.calcite.util.ImmutableIntList;
@@ -39,9 +38,9 @@ import org.apache.ignite.internal.cache.query.index.sorted.inline.SortedSegmente
 import org.apache.ignite.internal.cache.query.index.sorted.inline.io.InlineIO;
 import org.apache.ignite.internal.cache.query.index.sorted.keys.IndexKey;
 import org.apache.ignite.internal.cache.query.index.sorted.keys.IndexKeyFactory;
-import org.apache.ignite.internal.processors.cache.KeyCacheObject;
 import org.apache.ignite.internal.processors.cache.persistence.tree.BPlusTree;
 import org.apache.ignite.internal.processors.cache.persistence.tree.io.BPlusIO;
+import org.apache.ignite.internal.processors.cache.transactions.TransactionChanges;
 import org.apache.ignite.internal.processors.query.calcite.exec.exp.RangeIterable;
 import org.apache.ignite.internal.processors.query.calcite.exec.exp.TransformRangeIterable;
 import org.apache.ignite.internal.processors.query.calcite.schema.CacheTableDescriptor;
@@ -50,7 +49,6 @@ import org.apache.ignite.internal.processors.query.calcite.util.TypeUtils;
 import org.apache.ignite.internal.util.lang.GridCursor;
 import org.apache.ignite.internal.util.typedef.F;
 import org.apache.ignite.internal.util.typedef.internal.U;
-import org.apache.ignite.lang.IgniteBiTuple;
 import org.apache.ignite.spi.indexing.IndexingQueryFilter;
 import org.apache.ignite.spi.indexing.IndexingQueryFilterImpl;
 import org.jetbrains.annotations.Nullable;
@@ -81,7 +79,7 @@ public class IndexScan<Row> extends AbstractCacheColumnsScan<Row> {
      * First, set of keys changed (inserted, updated or removed) inside transaction: must be skiped during index scan.
      * Second, list of rows inserted or updated inside transaction: must be mixed with the scan results.
      */
-    private final IgniteBiTuple<Set<KeyCacheObject>, List<IndexRow>> txChanges;
+    private final TransactionChanges<IndexRow> txChanges;
 
     /**
      * @param ectx Execution context.
@@ -125,7 +123,7 @@ public class IndexScan<Row> extends AbstractCacheColumnsScan<Row> {
                 r -> new IndexRowImpl(rowHnd, r)
             );
 
-            txChanges.get2().sort(this::compare);
+            txChanges.newAndUpdatedEntries().sort(this::compare);
         }
         else
             txChanges = null;
@@ -257,7 +255,7 @@ public class IndexScan<Row> extends AbstractCacheColumnsScan<Row> {
 
         InlineIndexRowHandler rowHnd = idx.segment(0).rowHandler();
 
-        InlineIndexRowFactory rowFactory = (isInlineScan() && (txChanges == null || F.isEmpty(txChanges.get1()))) ?
+        InlineIndexRowFactory rowFactory = (isInlineScan() && (txChanges == null || F.isEmpty(txChanges.changedKeys()))) ?
             new InlineIndexRowFactory(rowHnd.inlineIndexKeyTypes().toArray(new InlineIndexKeyType[0]), rowHnd) : null;
 
         BPlusTree.TreeRowClosure<IndexRow, IndexRow> rowFilter = isInlineScan() ? null : createNotExpiredRowFilter();
@@ -429,9 +427,9 @@ public class IndexScan<Row> extends AbstractCacheColumnsScan<Row> {
                     new GridCursor[]{
                         // This call will change `txChanges.get1()` content.
                         // Removing found key from set more efficient so we break some rules here.
-                        new KeyFilteringCursor<>(idxCursor, txChanges.get1(), r -> r.cacheDataRow().key()),
+                        new KeyFilteringCursor<>(idxCursor, txChanges.changedKeys(), r -> r.cacheDataRow().key()),
                         new SortedListRangeCursor<>(
-                            IndexScan.this::compare, txChanges.get2(), lower, upper, lowerInclude, upperInclude)
+                            IndexScan.this::compare, txChanges.newAndUpdatedEntries(), lower, upper, lowerInclude, upperInclude)
                     },
                     idx.indexDefinition()
                 );
