+++ modules/calcite/src/main/java/org/apache/ignite/internal/processors/query/calcite/exec/TableScan.java
@@ -20,23 +20,20 @@ package org.apache.ignite.internal.processors.query.calcite.exec;
 import java.util.ArrayDeque;
 import java.util.Collections;
 import java.util.Iterator;
-import java.util.List;
 import java.util.NoSuchElementException;
 import java.util.Queue;
-import java.util.Set;
 import java.util.function.Function;
 import org.apache.calcite.util.ImmutableBitSet;
 import org.apache.ignite.IgniteCheckedException;
-import org.apache.ignite.internal.processors.cache.KeyCacheObject;
 import org.apache.ignite.internal.processors.cache.distributed.dht.topology.GridDhtLocalPartition;
 import org.apache.ignite.internal.processors.cache.persistence.CacheDataRow;
 import org.apache.ignite.internal.processors.cache.persistence.CacheSearchRow;
+import org.apache.ignite.internal.processors.cache.transactions.TransactionChanges;
 import org.apache.ignite.internal.processors.query.calcite.schema.CacheTableDescriptor;
 import org.apache.ignite.internal.util.lang.GridCursor;
 import org.apache.ignite.internal.util.lang.GridIteratorAdapter;
 import org.apache.ignite.internal.util.typedef.F;
 import org.apache.ignite.internal.util.typedef.internal.U;
-import org.apache.ignite.lang.IgniteBiTuple;
 import org.jetbrains.annotations.Nullable;
 
 /** */
@@ -70,7 +67,7 @@ public class TableScan<Row> extends AbstractCacheColumnsScan<Row> {
          * First, set of keys changed (inserted, updated or removed) inside transaction: must be skiped during index scan.
          * Second, list of rows inserted or updated inside transaction: must be mixed with the scan results.
          */
-        private IgniteBiTuple<Set<KeyCacheObject>, List<CacheDataRow>> txChanges;
+        private TransactionChanges<CacheDataRow> txChanges;
 
         /** */
         private Iterator<CacheDataRow> txIter = Collections.emptyIterator();
@@ -139,10 +136,10 @@ public class TableScan<Row> extends AbstractCacheColumnsScan<Row> {
                     if (txChanges != null) {
                         // This call will change `txChanges.get1()` content.
                         // Removing found key from set more efficient so we break some rules here.
-                        if (!F.isEmpty(txChanges.get1()))
-                            cur = new KeyFilteringCursor<>(cur, txChanges.get1(), CacheSearchRow::key);
+                        if (!F.isEmpty(txChanges.changedKeys()))
+                            cur = new KeyFilteringCursor<>(cur, txChanges.changedKeys(), CacheSearchRow::key);
 
-                        txIter = F.iterator0(txChanges.get2(), true, e -> e.key().partition() == part.id());
+                        txIter = F.iterator0(txChanges.newAndUpdatedEntries(), true, e -> e.key().partition() == part.id());
                     }
                 }
