+++ modules/calcite/src/test/java/org/apache/ignite/internal/processors/tx/SqlTransactionsUnsupportedModesTest.java
@@ -24,6 +24,7 @@ import org.apache.ignite.IgniteCheckedException;
 import org.apache.ignite.cache.query.SqlFieldsQuery;
 import org.apache.ignite.calcite.CalciteQueryEngineConfiguration;
 import org.apache.ignite.configuration.IgniteConfiguration;
+import org.apache.ignite.configuration.TransactionConfiguration;
 import org.apache.ignite.internal.IgniteEx;
 import org.apache.ignite.testframework.junits.common.GridCommonAbstractTest;
 import org.apache.ignite.transactions.Transaction;
@@ -53,7 +54,7 @@ public class SqlTransactionsUnsupportedModesTest extends GridCommonAbstractTest 
             sql("CREATE TABLE T1(ID BIGINT PRIMARY KEY, NAME VARCHAR(100))");
 
             for (TransactionConcurrency concurrency : TransactionConcurrency.values()) {
-                EnumSet<TransactionIsolation> supported = EnumSet.of(TransactionIsolation.READ_COMMITTED);
+                EnumSet<TransactionIsolation> supported = EnumSet.copyOf(TransactionConfiguration.TX_AWARE_QUERIES_SUPPORTED_MODES);
 
                 for (TransactionIsolation isolation : TransactionIsolation.values()) {
                     try (Transaction ignored = srv.transactions().txStart(concurrency, isolation)) {
